New code structure. Moved graphene clases to modules, split resolvers and models
