+++ gqlclans/message/models.py
@@ -1,0 +1,5 @@
+import graphene
+
+
+class Message(graphene.ObjectType):
+    body = graphene.String()
