+++ gqlclans/server_info/models.py
@@ -1,0 +1,6 @@
+import graphene
+
+
+class ServerInfo(graphene.ObjectType):
+    players_online = graphene.Int()
+    server = graphene.String()
